tests: Merge test_buffering into test_lower_exprs
